Ecommerce site/requirements.txt
flask
python-dotenv
groq
